validointi, l00gin, uutta M:ää, V:tä ja C:tä

+++ app/controllers/products_controller.php
@@ -3,7 +3,11 @@
 class ProductController extends BaseController {
     
     public static function index() {
-        $products = Product::all();
+        if (!parent::get_user_logged_in()) {
+            self::redirect_to('/login');
+        }
+        
+        $products = Product::find_by_owner(parent::get_user_logged_in()->id);
         $product_units = array();
         $product_departments = array();
         
@@ -17,18 +21,52 @@ class ProductController extends BaseController {
     }
     
     public static function show($id) {
-        $product = Product::find($id);
-        $product_unit = Unit::find($product->unit);
-        $product_department = Department::find($product->department);
+        $product = self::check_ownership_and_existence($id);
         
-        self::render_view('product/show.html', array('product' => $product, 
-            'product_unit' => $product_unit, 'product_department' => $product_department));
+        if (!$product) {
+            self::redirect_to('/products', array('message' => 'Tuotetta ei löytynyt!'));
+        } else {
+            $product_unit = Unit::find($product->unit);
+            $product_department = Department::find($product->department);
+
+            self::render_view('product/show.html', array('product' => $product, 
+                'product_unit' => $product_unit, 'product_department' => $product_department));
+        }
     }
     
-    public static function delete($id) {
-        Product::delete($id);
+    public static function edit($id) {
+        $product = self::check_ownership_and_existence($id);
         
-        self::redirect_to('/products', array('message' => 'Tuote poistettu!'));
+        if (!$product) {
+            self::redirect_to('/products', array('message' => 'Tuotetta ei löytynyt!'));
+        } else {
+            $units = Unit::all();
+            $departments = Department::all();
+
+            self::render_view('product/edit.html', array('attributes' => $product, 'units' => $units, 'departments' => $departments));
+        }
+    }
+    
+    public static function fave($id) {
+        $product = self::check_ownership_and_existence($id);
+        
+        if ($product) {
+            Product::fave($id, parent::get_user_logged_in()->id);
+        }
+        
+        self::redirect_to('/products');
+    }
+
+    public static function delete($id) {
+        $product = self::check_ownership_and_existence($id);
+        
+        if ($product) {
+            Product::delete($id);
+
+            self::redirect_to('/products', array('message' => 'Tuote poistettu!'));
+        } else {
+            self::redirect_to('/products', array('message' => 'Tuotetta ei löytynyt!'));
+        }
     }
     
     public static function create() {
@@ -40,14 +78,65 @@ class ProductController extends BaseController {
     
     public static function store() {
         $params = $_POST;
+        $units = Unit::all();
+        $departments = Department::all();
         
-        $id = Product::create(array(
+        $attributes = array(
             'name' => $params['name'],
             'department' => $params['department'],
             'unit' => $params['unit'],
-            'owner' => 1
-        ));
+            'owner' => parent::get_user_logged_in()->id
+        );
         
-        self::redirect_to('/product/' . $id, array('message' => 'Tuote lisätty!'));
+        $product = new Product($attributes);
+        $errors = $product->errors();
+        
+        if (count($errors) == 0) {
+            $id = Product::create($attributes);
+            self::redirect_to('/product/' . $id, array('message' => 'Tuote lisätty!'));
+        } else {
+            self::render_view('product/new.html', array('errors' => $errors, 'attributes' => $attributes, 'units' => $units, 'departments' => $departments));
+        }
+        
+    }
+    
+    public static function update($id) {
+        $params = $_POST;
+        $units = Unit::all();
+        $departments = Department::all();
+        $name_original = Product::find($id)->name;
+        
+        $attributes = array(
+            'name' => $params['name'],
+            'department' => $params['department'],
+            'unit' => $params['unit'],
+            'owner' => parent::get_user_logged_in()->id
+        );
+        
+        $product = new Product($attributes);
+        $errors = $product->errors();
+        
+        if (count($errors) == 0) {
+            Product::update($id, $attributes);
+            self::redirect_to('/product/' . $id, array('message' => 'Tuotetta on muokattu onnistuneesti!'));
+        } else {
+            $attributes['id'] = $id;
+            self::render_view('product/edit.html', array('errors' => $errors, 'attributes' => $attributes, 'units' => $units, 'departments' => $departments, 'name_original' => $name_original));
+        }
+    }
+    
+    public static function check_ownership_and_existence($product_id) {
+        $product = Product::find($product_id);
+        
+        if (!parent::get_user_logged_in()) {
+            return null;
+        }
+        
+        if ($product == null
+                || $product->owner != parent::get_user_logged_in()->id) {
+            return null;
+        } else {
+            return $product;
+        }
     }
 }
